qa-chatbot/pyproject.toml
[project]
name = "qa-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.302",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.1.0",
]
